Create ShareInspectorUI package for all SwiftUI views
I did this because the Xcode previews don't work in an app extension target. I can't get the previews to work with the separate UI package, either, but at least this potentially improves the separation of concerns, and will allow us to reuse UI components in the main app if needed.